# Conflicts:
#	mzmine-community/src/main/java/io/github/mzmine/gui/MZmineGUI.java

+++ mzmine-community/src/main/java/io/github/mzmine/datamodel/features/types/annotations/SmilesStructureType.java
@@ -61,7 +61,7 @@ public class SmilesStructureType extends StringType implements EditableColumnTyp
   }
 
   @Override
-  public int getPrefColumnWidth() {
+  public double getPrefColumnWidth() {
     return 100;
   }
 }
